factor LibraryButton into own file component & fix mobile
