+++ src/components/LayerUI.tsx
@@ -47,6 +47,7 @@ import { Tooltip } from "./Tooltip";
 import { UserList } from "./UserList";
 import Library from "../data/library";
 import { JSONExportDialog } from "./JSONExportDialog";
+import { LibraryButton } from "./LibraryButton";
 
 interface LayerUIProps {
   actionManager: ActionManager;
@@ -547,44 +548,6 @@ const LayerUI = ({
     />
   ) : null;
 
-  const LibraryIcon = () => {
-    const LIBRARY_ICON = (
-      <svg viewBox="0 0 512 512">
-        <path
-          fill="currentColor"
-          d="M116.65 219.35a15.68 15.68 0 0 0 22.65 0l96.75-99.83c28.15-29 26.5-77.1-4.91-103.88C203.75-7.7 163-3.5 137.86 22.44L128 32.58l-9.85-10.14C93.05-3.5 52.25-7.7 24.86 15.64c-31.41 26.78-33 74.85-5 103.88zm143.92 100.49h-48l-7.08-14.24a27.39 27.39 0 0 0-25.66-17.78h-71.71a27.39 27.39 0 0 0-25.66 17.78l-7 14.24h-48A27.45 27.45 0 0 0 0 347.3v137.25A27.44 27.44 0 0 0 27.43 512h233.14A27.45 27.45 0 0 0 288 484.55V347.3a27.45 27.45 0 0 0-27.43-27.46zM144 468a52 52 0 1 1 52-52 52 52 0 0 1-52 52zm355.4-115.9h-60.58l22.36-50.75c2.1-6.65-3.93-13.21-12.18-13.21h-75.59c-6.3 0-11.66 3.9-12.5 9.1l-16.8 106.93c-1 6.3 4.88 11.89 12.5 11.89h62.31l-24.2 83c-1.89 6.65 4.2 12.9 12.23 12.9a13.26 13.26 0 0 0 10.92-5.25l92.4-138.91c4.88-6.91-1.16-15.7-10.87-15.7zM478.08.33L329.51 23.17C314.87 25.42 304 38.92 304 54.83V161.6a83.25 83.25 0 0 0-16-1.7c-35.35 0-64 21.48-64 48s28.65 48 64 48c35.2 0 63.73-21.32 64-47.66V99.66l112-17.22v47.18a83.25 83.25 0 0 0-16-1.7c-35.35 0-64 21.48-64 48s28.65 48 64 48c35.2 0 63.73-21.32 64-47.66V32c0-19.48-16-34.42-33.92-31.67z"
-        />
-      </svg>
-    );
-
-    return (
-      <label
-        className={clsx(
-          "ToolIcon  ToolIcon_type_floating zen-mode-visibility",
-          `ToolIcon_size_m`,
-          {
-            "zen-mode-visibility--hidden": appState.zenModeEnabled,
-          },
-        )}
-        title={`${capitalizeString(t("toolBar.library"))} — 9`}
-        style={{ marginInlineStart: "var(--space-factor)" }}
-      >
-        <input
-          className="ToolIcon_type_checkbox"
-          type="checkbox"
-          name="editor-library"
-          onChange={(event) => {
-            setAppState({ isLibraryOpen: event.target.checked });
-          }}
-          checked={appState.isLibraryOpen}
-          aria-label={capitalizeString(t("toolBar.library"))}
-          aria-keyshortcuts="9"
-        />
-        <div className="ToolIcon__icon">{LIBRARY_ICON}</div>
-      </label>
-    );
-  };
-
   const renderFixedSideContainer = () => {
     const shouldRenderSelectedShapeActions = showSelectedShapeActions(
       appState,
@@ -628,7 +591,10 @@ const LayerUI = ({
                         />
                       </Stack.Row>
                     </Island>
-                    <LibraryIcon />
+                    <LibraryButton
+                      appState={appState}
+                      setAppState={setAppState}
+                    />
                   </Stack.Row>
                   {libraryMenu}
                 </Stack.Col>
